master plan

+++ shoppino/server/src/main/java/com/camino/camino_v1/service/RefreshTokenService.java
@@ -4,12 +4,14 @@ import java.time.Instant;
 import java.util.Optional;
 import java.util.UUID;
 
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.camino.camino_v1.model.RefreshToken;
+import com.camino.camino_v1.model.User;
 import com.camino.camino_v1.repository.RefreshTokenRepository;
 import com.camino.camino_v1.repository.UserRepository;
 import com.camino.camino_v1.util.AppConstants;
@@ -28,16 +30,29 @@ public class RefreshTokenService {
 
     Logger logger = LoggerFactory.getLogger(RefreshTokenService.class);
 
-    public RefreshToken createRefreshToken(String username) {
+    public RefreshToken createRefreshToken(long user_id) {
 
-        RefreshToken refreshToken = RefreshToken.builder()
-            .username(username)
+        Optional<User> optionalUser = userRepository.findById(user_id);
+        if(optionalUser.isPresent()) {
+         RefreshToken refreshToken = RefreshToken.builder()
+            .user(optionalUser.get())
             .token(UUID.randomUUID().toString())
             .expiryDate(Instant.now().plusMillis(appConstants.getJWT_REFRESHTOKEN_EXPIRATION_TIME()))
             .build();
 
-        logger.info("Createing a new RefreshToken {}", refreshToken.toString());
-        return refreshTokenRepository.save(refreshToken);
+        logger.info("Createing a new RefreshToken {} \n For {}", refreshToken.toString(), optionalUser.get().toString());   
+        try {
+            return refreshTokenRepository.save(refreshToken);
+        } catch (Exception e) {
+            throw new Error("during refresh t creation "+e.getMessage());
+        }
+       } else {
+        throw new Error("Null user For refresh token creation");
+       }
+
+        
+
+        
     }
 
     public Optional<RefreshToken> findByToken(String token) {
@@ -47,17 +62,21 @@ public class RefreshTokenService {
     public RefreshToken varifyExpiration(String token) {
         RefreshToken refreshToken = FindTokenFromDataBase(token);
 
-            if(refreshToken.getExpiryDate() == null) {
-                logger.info("Attend varifyExpiration of refreshtoken , and the expiryData is Null {}", refreshToken.getToken());
-                return null;
-            }            
+            if(refreshToken != null) {
 
-        if(refreshToken.getExpiryDate().isBefore(Instant.now())) {
-            refreshTokenRepository.delete(refreshToken);
-            logger.error("Expiraed refreshToken Deleted {}", refreshToken.toString());
+                if(refreshToken.getExpiryDate() == null) {
+                    logger.info("Attend varifyExpiration of refreshtoken , and the expiryData is Null {}", refreshToken.getToken());
+                    return null;
+                }            
+                if(refreshToken.getExpiryDate().isBefore(Instant.now())) {
+                    refreshTokenRepository.delete(refreshToken);
+                    logger.error("Expiraed refreshToken Deleted {}", refreshToken.toString());
+                    return null;
+                } 
+                return refreshToken;
+            }
             return null;
-        } 
-        return refreshToken;
+
     }
 
     private RefreshToken FindTokenFromDataBase(String token) {
@@ -71,6 +90,6 @@ public class RefreshTokenService {
 
     public RefreshToken reGenerate(RefreshToken refreshToken) {
         refreshTokenRepository.delete(refreshToken);
-        return createRefreshToken(refreshToken.getUsername());
+        return createRefreshToken(refreshToken.getUser().getId());
     }
 }
